Linear mixing: began converting linear_mix to take a list
The init function now takes a list of models and a list specifying the
number of free parameters for each model.

Work is being done to calculate the mixed loglikelihood, but this
requires `utils.mixing_function` to take as method `dirichlet` or other
multivariate distributions. This will be the next step.