Global linear mixing: Seems to work now!!
Threre is one important observation that had to be made to make this
work.
Namely that if $X$ is a lognormal distributed variable, then $\log X$ is
a normal distributed random variable, not the other way around out.
So I assumed that the parameters of the Dirichlet distribution are
lognormal distirbuted, and so the log of these parameters is normall
distributed.

An example where this works has been made but not converted into a
notebook yet for the website.
These details will be forth coming