(Issue #82) Have trees & tests get mpirun command from openbtmixing.
In the 1_DiffBuilds branch in my openbtmixing clone, I used the new build
script to build a wheel for macOS/arm64.  Used delocate to add in all external
library dependencies and reset all RPATHs in package to those libraries.
Installed for use with Taweret, made updates in the commit, and see all tests
pass.