(Issue #91) tox flake8 check already checks all python code.
Apparently installing the python package in the root of the repo and with the
tox.ini in the root implies that flake8 will automatically check all Python
code in the repo including the tools folder, which is not really a part of the
package.  Therefore, we don't need the separate check script.