Print info to debug new GH action/macOS failure.
This associated action is failing for macOS 3.12/Python 3.10 at the very least
because it cannot find the XCode command line tools.

git: error: unable to locate xcodebuild, please make sure the path to the Xcode
folder is set correctly!

I have verified that XCode and the CLT are supposed to be installed for that
runner and this wasn't happening earlier.