(Issue #91) Intentionally alter docs to confirm immediate publishing to RTD.
This is to confirm that the docs generation/publishing is setup so that any
pushes to main trigger an automatic publishing to RTD.  In particular, this is
accomplished outside of GH actions and special branches.  This is also a
confirmation that hot fixes could be made to the docs if so desired.

THESE CHANGES SHOULD BE REVERTED DIRECTLY ON MAIN.