(Issue #91) Intentionally alter docs.
Made additions to .rst and python docstring docs to see that they are rendered
locally by tox, that they are visible in the RTD test rendering for the PR, but
not published in the official, public RTD docs.  They did show up as expected
with tox.

NOTE:  I want these to make it into main so that we do see them automatically
incluced in the official, public RTD docs.  We can then remove them as a hotfix
to test the whole workflow.  Note that these were also added in to generate a
merge conflict with the changes made direclty on main at commit d53cbfe so that
we can test the git workflow a bit as well.