(Issue #91) Run PR actions less.
Some were running on edits to PR conversations as well as upon closing.  Revert
to default, which seems to be working for other PR actions.  If we want actions
to run upon merging, then we will make actions on special branches in accord
with the git workflow to run on pushes to the branch.