Revert action to nominal setup after debug run.
That test revealed that macos-14 is ARM64 and that the package tests on that
runner are pulling down the openbtmixing-1.0.1-py3-none-macosx_12_0_arm64.whl
wheel.  While it runs through successfully, the otool output shows that the
OpenBT CLT libraries are set to the installation path on my machine at the time
of creating the wheel.  Not clear how the tests run through.